adding option -fdelete-pkg to -O2 for can_delete()

+++ lib/B/C.pm
@@ -266,7 +266,7 @@ my %all_bc_pkg = map {$_=>1}
      AnyDBM_File Fcntl Regexp overload Errno Exporter Exporter::Heavy Config
      warnings warnings::register DB next maybe maybe::next FileHandle fields vars
      AutoLoader Carp Symbol PerlIO PerlIO::scalar SelectSaver ExtUtils ExtUtils::Constant
-     ExtUtils::Constant::ProxySubs threads base
+     ExtUtils::Constant::ProxySubs threads base IO::File
     );
 # B::C stash footprint: mainly caused by blib, warnings, and Carp loaded with DynaLoader
 # perl5.15.7d-nt -MO=C,-o/dev/null -MO=Stash -e0
@@ -295,7 +295,7 @@ our ($module, $init_name, %savINC, $mainfile);
 our ($use_av_undef_speedup, $use_svpop_speedup) = (1, 1);
 our ($pv_copy_on_grow, $optimize_ppaddr, $optimize_warn_sv, $use_perl_script_name,
     $save_data_fh, $save_sig, $optimize_cop, $av_init, $av_init2, $ro_inc, $destruct,
-    $fold, $warnings, $const_strings, $stash);
+    $fold, $warnings, $const_strings, $stash, $can_delete_pkg);
 our $verbose = 0;
 our %option_map = (
     'cog'             => \$B::C::pv_copy_on_grow,
@@ -305,6 +305,7 @@ our %option_map = (
     'warn-sv'         => \$B::C::optimize_warn_sv,
     'av-init'         => \$B::C::av_init,
     'av-init2'        => \$B::C::av_init2,
+    'delete-pkg'      => \$B::C::can_delete_pkg,
     'ro-inc'          => \$B::C::ro_inc,
     'stash'           => \$B::C::stash,    # disable with -fno-stash
     'destruct'        => \$B::C::destruct, # disable with -fno-destruct
@@ -4950,6 +4951,12 @@ sub skip_pkg {
   return 0;
 }
 
+# with -O0 or -O1 do not delete packages which were brought in from
+# the script, i.e. not defined in B::C or O. Just to be on the safe side.
+sub can_delete {
+  return $can_delete_pkg or $all_bc_pkg{$_{0}};
+}
+
 sub should_save {
   no strict qw(vars refs);
   my $package = shift;
@@ -5006,7 +5013,7 @@ sub should_save {
   # Omit the packages which we use (and which cause grief
   # because of fancy "goto &$AUTOLOAD" stuff).
   # XXX Surely there must be a nicer way to do this.
-  if (skip_pkg($package)) {
+  if (skip_pkg($package) and can_delete($package) ) {
     delete_unsaved_hashINC($package);
     return; # $include_package{$package} = 0;
   }
@@ -5019,7 +5026,9 @@ sub should_save {
         warn "Cached $package is already deleted\n";
       }
     }
-    delete_unsaved_hashINC($package) unless $include_package{$package};
+    if (!$include_package{$package} and can_delete($package)) {
+      delete_unsaved_hashINC($package);
+    }
     return $include_package{$package};
   }
 
@@ -5039,7 +5048,9 @@ sub should_save {
       return mark_package($package);
     }
   }
-  delete_unsaved_hashINC($package) unless $package =~ /^PerlIO/;
+  if ($package !~ /^PerlIO/ and can_delete($package)) {
+    delete_unsaved_hashINC($package);
+  }
   return $include_package{$package} = 0;
 }
 
@@ -5520,7 +5531,7 @@ sub compile {
   my %optimization_map = (
     0 => [qw()],                # special case
     1 => [qw(-fcog -fppaddr -fwarn-sv -fav-init2)], # falls back to -fav-init
-    2 => [qw(-fro-inc -fsave-data)],
+    2 => [qw(-fro-inc -fsave-data -fdelete-pkg)],
     3 => [qw(-fno-destruct -fconst-strings -fno-fold -fno-warnings)],
     4 => [qw(-fcop)],
   );
@@ -5667,6 +5678,7 @@ OPTION:
         push @opt, @{ $optimization_map{$i} }
           if exists $optimization_map{$i};
       }
+      $opt{$arg}++;
       unshift @options, @opt;
       warn "options : ".(join " ",@opt)."\n" if $verbose;
     }
@@ -5933,6 +5945,15 @@ enabled automatically where it is known to work.
 
 Enabled with C<-O2>.
 
+=item B<-fdelete-pkg>
+
+Delete packages which appear to be nowhere used automatically.
+This creates smaller executables but might miss run-time called
+methods. Note that you can always use -u to add automatically
+deleted packages.
+
+Enabled with C<-O2>.
+
 =item B<-fconst-strings>
 
 Declares readonly strings as const. Enables C<-fcog>.
@@ -6027,7 +6048,7 @@ Note that C<-fcog> without C<-fno-destruct> will be disabled >= 5.10.
 
 =item B<-O2>
 
-Enable B<-O1> plus B<-fro-inc> and B<-fsave-data>.
+Enable B<-O1> plus B<-fro-inc>, B<-fsave-data> and B<-fdelete-pkg>.
 
 =item B<-O3>
 
@@ -6035,7 +6056,8 @@ Enable B<-O2> plus B<-fno-destruct> and B<-fconst-strings>.
 
 =item B<-O4>
 
-Enable B<-O3> plus B<-fcop>. Very unsafe, 10% faster, 10% smaller.
+Enable B<-O3> plus B<-fcop>. Very unsafe, rarely works,
+10% faster, 10% smaller.
 
 =back
